K8s-configMap/pyproject.toml
[project]
name = "k8s-configmap"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.128.2",
    "python-decouple>=3.8",
    "uvicorn>=0.40.0",
]
